requirements.txt
faiss-cpu==1.8.0.post1
flask==3.0.3
langchain==0.2.17
langchain-community==0.2.19
langchain-huggingface==0.0.3
langchain-openai==0.1.25
pypdf==4.3.1
python-dotenv==1.0.1
sentence-transformers==3.0.1
